Annotation update

+++ annotations/nocedit.exe/pseudocode/src/engine/3d.c/renderOverlayTextureEnable_FUN_00404020.c
@@ -38,7 +38,7 @@ void * __cdecl engine_3d_c_renderOverlayTextureEnable_FUN_00404020(SMRGLHeaderPr
     *piVar3 = (pSVar2->base).type + g_ProcessedVertexOffset;
     g_RenderVertexBuffer[(pSVar2->base).type + uVar1].u = (pSVar2->base).count;
     piVar3 = piVar3 + 1;
-    g_RenderVertexBuffer[(pSVar2->base).type + uVar1].v = (pSVar2->surface_normal).A;
+    g_RenderVertexBuffer[(pSVar2->base).type + uVar1].v = (int)(pSVar2->surface_normal).A;
     pSVar2 = (SMRGLHeaderPrimitive *)&(pSVar2->surface_normal).B;
   }
   engine_clipper_c_clipAndRasterize_FUN_004371b0((prim->base).count,g_ProcessedVertexIndices);
