+++ annotations/nocedit.exe/pseudocode/src/engine/3d.c/renderPrimitivePlaneMaskedComplex_FUN_004046c0.c
@@ -51,7 +51,7 @@ LAB_004047c8:
     g_RenderVertexBuffer[(pSVar2->base).type + g_ProcessedVertexOffset].u = (pSVar2->base).count;
     piVar3 = piVar3 + 1;
     g_RenderVertexBuffer[(pSVar2->base).type + g_ProcessedVertexOffset].v =
-         (pSVar2->surface_normal).A;
+         (int)(pSVar2->surface_normal).A;
     pSVar2 = (SMRGLHeaderPrimitive *)&(pSVar2->surface_normal).B;
   }
   engine_clipper_c_clipAndRasterize_FUN_004371b0((primitive->base).count,g_ProcessedVertexIndices);
