+++ annotations/nocedit.exe/pseudocode/src/engine/drender.cpp/CDemonRenderer_renderPremium_FUN_0048b1e0.cpp
@@ -28,7 +28,7 @@ void __cdecl engine_drender_cpp_CDemonRenderer_renderPremium_FUN_0048b1e0(CDemon
           iVar1 = pSVar3[1].base.type;
           if (this_ptr->skip_uv_extraction == 0) {
             this_ptr->vertex_buffer_ptr[iVar1].u = pSVar3[1].base.count;
-            this_ptr->vertex_buffer_ptr[iVar1].v = pSVar3[1].surface_normal.A;
+            this_ptr->vertex_buffer_ptr[iVar1].v = (int)pSVar3[1].surface_normal.A;
           }
           *(int *)((int)g_VertexIndexBuffer + iVar5) = iVar1;
           pSVar3 = (SMRGLHeaderPrimitive *)&(pSVar3->surface_normal).B;
